pyproject.toml
[tool.poetry]
name = "rag_chat"
version = "0.1.0"
description = ""
authors = ["Alex Bru <brubuxo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
langchain = ">=0.0.346"
pymongo = {version = ">=4.4", extras = ["srv"]}
python-dotenv = "^1.0.0"
pandas = "^2.1.4"
llama-index = "^0.9.30"
ipykernel = "^6.28.0"
qdrant-client = "^1.7.0"
pydantic = "^2.5.3"
openai = "^1.11.1"
motor = "^3.3.2"
langchainhub = "^0.1.14"
langchain-openai = "^0.0.5"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.28.0"
missingno = "^0.5.2"
anytree = "^2.12.1"
spacy = "^3.7.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
